+++ crates/sncast/src/starknet_commands/balance.rs
